exact/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

confusable-homoglyphs==3.2.0
drf-flex-fields==1.0.2
django-appconf==1.0.5
django-extensions==3.2.3
django-filter==23.2
django-friendly-tag-loader==1.3.1
django-project-version==0.18.1
django-redis==5.3.0
django-registration==3.4
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
uritemplate==3.0.1
fasteners==0.15
monotonic==1.5
Pillow==10.0.0
psycopg2-binary==2.9.6
pytz==2023.3
gunicorn==21.2.0
opencv-python==4.8.0.74
openslide-python==1.3.0
tifffile>=2023.7.0
pyvips==2.2.1
pydicom==2.4.2
ptvsd==4.3.2
czifile==2019.7.2
gitpython==3.1.11
gdown==4.7.1
pandas==2.0.3
qt-wsi-registration>=0.0.6
EXCAT-Sync>=0.0.38
pyyaml>=6.0.1
aicsimageio==4.11.0
